fix: improve menu expanding animation
The side-menu has two states, one narower with only icons and other wider with both icons and text.
Previously when expanding the menu each item of the menu wrapped into two lines during the animation.
This made the transition look inconsistent.

To mitigate this, instead of conditionally rendering the text part it was renderd and hide under it's parent.
Also menu item containers were changed to inline-flex and width of each menu item was made fix.